requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
youtube-transcript-api
google-generativeai
faiss-cpu
